requirements.txt
selenium
webdriver-manager
pandas
python-dotenv
python-telegram-bot
apscheduler
requests
